requirements.txt
amqp==5.3.1
anaconda-anon-usage @ file:///home/task_176160626771849/conda-bld/anaconda-anon-usage_1761606317374/work
anaconda-auth @ file:///C:/miniconda3/conda-bld/anaconda-cloud-auth-split_1759943207790/work
anaconda-cli-base @ file:///C:/miniconda3/conda-bld/anaconda-cli-base_1762295036999/work
annotated-types @ file:///C:/miniconda3/conda-bld/annotated-types_1761745107361/work
anyio==4.12.0
arabic-reshaper==3.0.0
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.1
boltons @ file:///C:/b/abs_e2_iokhxbp/croot/boltons_1751383740243/work
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1761745219545/work
cachetools==6.2.4
cbor2==5.7.1
celery==5.4.0
certifi @ file:///C:/miniconda3/conda-bld/certifi_1759786076808/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832792955/work
channels==4.0.0
channels-redis==4.2.0
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761744975868/work
click @ file:///C:/b/abs_42rbjngfys/croot/click_1744271606983/work
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1729036581634/work
conda @ file:///C:/miniconda3/conda-bld/conda_1760457894461/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///C:/b/abs_c9yejtsx9g/croot/conda-anaconda-tos_1755123332641/work
conda-content-trust @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-content-trust_1729088072778/work
conda-libmamba-solver @ file:///opt/miniconda3/conda-bld/conda-libmamba-solver_1761744809966/work/src
conda-package-handling @ file:///C:/miniconda3/conda-bld/conda-package-handling_1762366515145/work
conda_package_streaming @ file:///C:/miniconda3/conda-bld/conda-package-streaming_1762361689679/work
constantly==23.10.4
coverage==7.13.0
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.1.0
Deprecated==1.3.1
distro @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/distro_1729059153117/work
Django==5.2.9
django-cors-headers==4.3.1
django-csp==4.0
django-encrypted-model-fields==0.6.5
django-filter==23.5
django-htmx==1.18.0
django-otp==1.5.4
django-ratelimit==4.1.0
django-redis==5.4.0
django-tenants==3.9.0
django-tinymce==4.1.0
django-unfold==0.42.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
exceptiongroup==1.3.1
execnet==2.1.2
factory_boy==3.3.3
Faker==39.0.0
freetype-py==2.5.1
freezegun==1.5.5
frozendict @ file:///C:/miniconda3/conda-bld/frozendict_1761750728192/work
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==22.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperlink==21.0.0
idna @ file:///C:/miniconda3/conda-bld/idna_1761912043388/work
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
jaraco.classes @ file:///C:/b/abs_6erueoob1v/croot/jaraco.classes_1755516340851/work
jaraco.context @ file:///C:/Users/dev-admin/buildout/perseverance-python-buildout/croot/jaraco.context_1731721000658/work
jaraco.functools @ file:///C:/b/abs_a1jv6v_pzp/croot/jaraco.functools_1740408129620/work
jsonpatch @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/jsonpatch_1729054776004/work
jsonpointer @ file:///C:/b/abs_73u73l7pl9/croot/jsonpointer_1753788460913/work
keyring @ file:///C:/b/abs_596d81ux2c/croot/keyring_1745484709696/work
kombu==5.4.2
libmambapy @ file:///C:/miniconda3/conda-bld/mamba-split_1757623139479/work/libmambapy
lxml==6.0.2
markdown-it-py @ file:///C:/b/abs_3espcrdl35/croot/markdown-it-py_1756299213703/work
mdurl @ file:///C:/miniconda3/conda-bld/mdurl_1758552277768/work
menuinst @ file:///C:/miniconda3/conda-bld/menuinst_1762860868009/work
more-itertools @ file:///C:/miniconda3/conda-bld/more-itertools_1761121564395/work
msgpack==1.1.2
numpy==2.4.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.0.0
pkce @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pkce_1729049216383/work
platformdirs @ file:///C:/miniconda3/conda-bld/platformdirs_1762356623609/work
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.0
psycopg2-binary==2.9.10
py-ubjson==0.16.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496153123/work
pydantic @ file:///C:/miniconda3/conda-bld/pydantic_1762505416722/work
pydantic-settings @ file:///C:/miniconda3/conda-bld/pydantic-settings_1758784270277/work
pydantic_core @ file:///C:/miniconda3/conda-bld/pydantic-core_1760597457096/work
Pygments @ file:///C:/miniconda3/conda-bld/pygments_1762431428918/work
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT @ file:///C:/Users/dev-admin/buildout/croot/pyjwt_1739559988030/work
pyOpenSSL==25.3.0
pyotp==2.9.0
pyparsing==3.3.1
pypdf==6.5.0
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753030965/work
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-xdist==3.8.0
python-barcode==0.16.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==4.0.0
pytz==2025.2
pywin32-ctypes @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pywin32-ctypes_1729046491215/work
PyYAML==6.0.3
qrcode==8.0
readchar @ file:///C:/miniconda3/conda-bld/readchar_1760613474723/work
redis==5.2.1
reportlab==4.4.7
requests==2.32.3
rich @ file:///C:/miniconda3/conda-bld/rich_1760375661587/work
rlPyCairo==0.4.0
rsa==4.9.1
ruamel.yaml @ file:///C:/miniconda3/conda-bld/ruamel.yaml_1762536064547/work
ruamel.yaml.clib @ file:///C:/miniconda3/conda-bld/ruamel.yaml.clib_1762530094515/work
semver @ file:///C:/miniconda3/conda-bld/semver_1761903323755/work
sentry-sdk==2.48.0
service-identity==24.2.0
setuptools==80.9.0
shellingham @ file:///C:/miniconda3/conda-bld/shellingham_1761912227081/work
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.3
stripe==11.3.0
svglib==1.6.0
tenacity==9.1.2
tinycss2==1.5.1
tomli @ file:///C:/b/abs_88e598m6o8/croot/tomli_1753774604115/work
tqdm @ file:///C:/miniconda3/conda-bld/tqdm_1762863407729/work
truststore @ file:///C:/miniconda3/conda-bld/truststore_1762521027919/work
Twisted==25.5.0
txaio==25.12.2
typer @ file:///C:/miniconda3/conda-bld/typer_1757580777256/work
typing-inspection @ file:///C:/miniconda3/conda-bld/typing-inspection_1760614188477/work
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata==2024.2
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
uritools==6.0.1
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746278300/work
wrapt==2.0.1
xhtml2pdf==0.2.17
zope.interface==8.1.1
zstandard @ file:///C:/miniconda3/conda-bld/zstandard_1758189089298/work
